+++ CHANGELOG.rst
@@ -4,9 +4,7 @@ Changelog
 
 Unreleased
 ==========
-
-1.0.3 (2022-03-09)
-==================
+* feat: Expand/collapse ALL nodes for navigation tree in admin changelist view.
 * feat: Expand/collapse enabled for navigation tree in admin changelist view.
 
 1.0.2 (2022-03-03)