+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 
 Unreleased
 ==========
+* feat: Expand/collapse enabled for navigation tree in admin changelist view.
 
 1.0.2 (2022-03-03)
 ==================