+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 
 Unreleased
 ==========
+* fix: Extended app Node deletion 404 bugfix
 
 1.0.2 (2022-03-03)
 ==================