+++ CHANGELOG.rst
@@ -4,7 +4,8 @@ Changelog
 
 Unreleased
 ==========
-* Added delete confirmation template to overwrite delete view breadcrumbs
+* feat: Added delete confirmation template to overwrite delete view breadcrumbs
+* fix: Extended app Node deletion 404 bugfix
 
 1.0.2 (2022-03-03)
 ==================