+++ CHANGELOG.rst
@@ -5,6 +5,8 @@ Changelog
 Unreleased
 ==========
 * feat: Expand/collapse enabled for navigation tree in admin changelist view.
+* feat: Added delete confirmation template to overwrite delete view breadcrumbs
+* fix: Extended app Node deletion 404 bugfix
 
 1.0.2 (2022-03-03)
 ==================